Add files via upload
Changed all functions that user will interact with, so they don't need to give parameters such as data, variable_metadata, QC_samples and so on. All those will be set once at the begining and saved up into the object. The functions themselves thus require much less parameters for user to put it or understand. Which should streamline the usage